POC: Daydream with alt placeholder
Idea:
Add a daydream that will show art or whatever while music is playing
but allow an alternate dream to show while music isnt playing,

Implementation:
In order to attach to the DreamManagerService we must declare
WRITE_DREAM but that is a signature protected permission so its out
Instead we can cheat, once our dream is started the DreamController
passes us a window token which is just a raw IBinder, using
reflection we can pull that token from our super class,
then we simply bind with any available dream
(android requires? all dream services to be exported
so this is no problem). and forward our token to the alt
dream. letting it draw on top of us

Ive also tested and we can query for available dreams
to allow the user to choose any dream installed.

Change-Id: I89972e8cb207c050a361b95a54c0212f876b126d