initial changes to combined queue
in order for local and remote songs to mingle they need
to share common whatevers, trying to pass the song objects
directly to the service was ridiculously slow, so trying a different
approach. the idea is that any song local or remote that is
sent to the service must first be entered into the local database
the service then uses the local db under the existing cursor
model to reference songs, this is much faster than pacelizing
the list, and will eventuall enable after the fact metadata parsing
for "dumb" remote libraries that contain no metadata.

Change-Id: Ice9c0e89b9dc63ceef2287adb51e9414a0a2e26c