core-artwork: setup for its own process
split requestmanager into a requestor and a fetcher
the requestor will run in the ui process and query
the artwork provider if what it needs isnt in its cache
the artwork provider will check the diskcache then
send a request to the fetcher to retrieve it, the fetcher
will save the bitmap to the diskcache and notify on the uri
used for the request.  then the requestor
will requery the artwork provider to retrieve the image from
the diskcache. (that last parts not done yet)

this should dramatically reduce the memory usage of the ui
process when large amounts of images are missing as well as
prevent the ui process from being restarted when the playback
service needs artwork

Change-Id: Ia8c0c56372735592a893767ca599b490454ecb41